batching execution ign 33 (#37)
The batching algo and its design is covered in more detail under: `./docs/design.md`

Execution is now batched based on the dependencies of the execution graph. Each batch is formed by asking if there are unstarted nodes that have all there dependencies completed. A batch is then executed in parallel and run to completion. The results of the batch execution are then merged into the execution state and the deployment completed or a new batch formed. An error stops the execution after all other vertexes in the batch have completed/errored.

The state management was updated across the app to better support the UI while execution. Now a Deployment class is used to capture the state associated with the deployment process. This is now threaded through execution graph creation and execution. As each step in the process completes, the result is merged with the Deployment state, and the ui is updated.

The batcher similarly manipulates the Deployment state to determine the next batch, then executes it. Each vertex of the batch is triggered at the same time. The UI is updated on each vertex completing, but the batch is only complete when all vertexes in the batch are complete.

The ink.js based UI has been moved from core to hardhat-plugin, with ignition now accepting a render func that takes the Deployment state as an arg and is triggered when the Deployment state is updated.