Added other Hardhat tutorials
Hey there - love this tutorial as a way to get started using HardHat. I work for Alchemy and our developers tell us they find this library of content useful when they're getting started. One thing is that they've mentioned they also wanted to know how to interact with the smart contract once it's deployed and how to submit it to Etherscan to log. As such, we created these tutorials using HardHat to address this need and they've been some of our top views. Adding them here in the tutorial closing section in case HardHat readers also had the same questions and find it useful as they are natural next steps. Lmk of any questions / any better place to put it, thanks!