feat:restart-deployment (#70)
Support restarting or rerunning deployments by recording a deployment run to a journal, then rebuilding the state from the journal on subsequent runs.

The journal is stored as a sibling to the js module file. It records the state change commands that get passed during the execution phase. On a rerun the journal is read and the execution state reconstructed. Any on hold or errored vertexes are set back to unstarted, then the execution runs again.

In parts:

* feat: add state command journal

Every command that updates the Deployment state, _during execution_, is
written to a Journal.

The journal is passed in at `Ignition` creation. In the
**hardhat-plugin**, the journal implementation writes to a file that is
a sibling to the module being deployed.

Supporting refactorings:

* remove `ui` flag infavour of constructor

The `Ignition` object can be setup with a UI renderer. If none is given
a noop renderer is put in its place. The deploy function can then focus
on `deploy` options.

This means the `IgnitionWrapper` creates a new ignition object in each
of its tasks.

* refactor(deploymentState): pull out run logic

Create a shared run command function to dry up the repetition on
conversion to commands.

* feat: read messages from journal

The journal uses ndjson, so read them one line at a time through an
async generator.

test(stateReducer): add tests and simplify

Increase the test coverage over the state reducer before changing the
message structure.

Supporting refactors:

* move reducer types out

* separate out execution state reducer

* deployState - restructure execution state

Use a map of vertex ids to states rather than collections of states.

The intermediary batch results are stored against the vertexes structure
rather than the batch. The batch is reduced down to a list of vertex
ids.

* deployState - remove end batch message

Replace the end batch message updating the state with a check in the
individual vertex update to see if the batch is complete.

We also change `batch` in execution state so that it is nullable, either
befor or after a batch is started/completed. This is a change from the
empty set that was used before.

* change the names of execution messages

* feat: show message if nothing to be rerun

If the journal indicates that all contract deploys/calls have already
completed, then display a message to that effect at the cli and then
display the previously deployed contract addresses.

Supporting refactor:

* refactor(core): merge batch into execution

In the short term there is no other planned execution policy other than
batching so merge the batching code into execution proper.

* feat: add hold state

Allow execution on a vertex to return a `hold` result.

* chore: run build on tests

Enforce typing in tests by running `tsc` on tests. This is done through
the cli and included in the `fullcheck` script.

* feat(core): reach end on hold

Either a failure or a on-hold from a vertex stops further execution. The
deployment is complete.

* feat(cli): show message on hold

Display more details of the on hold vertexes when a deployment stops as
on hold.

Supporting refactors:

* refactor(core): remove journal from tx service

The journaling is happening at the execution graph level for the moment.

* refactor(core): encapsulate providers in services

Seal the usage of providers behind services to ease mocking.

* refactor: externalise services

Move services to be external to ignition core. This is to support
mocking. We should look again at the best api for this.

* test(core): add rerun tests

Component level testing of reruns using an in-memory journal and mocking
out the services.

* fix(plugin): serialization/deserializtion of messages

Support more complex return results from ethers in the messages for the
journal.

We need to be able to serialize/deserialize results with BigNumber and
complex `Error` objects. For now we are only holding on to the message
in the error.

refactor(core): remove old journal files

* chore: updating multisig example

Change the multisig example so that an external script is required.
Reflect an equivalent of that in the matching test.